improved makefile

+++ upload_rates.py
@@ -1,6 +1,6 @@
 from mp_functions import upload_rates
 import sys
-unix_time=float(sys.argv[1])
+unix_time=int(sys.argv[1])
 base=sys.argv[2]
 quote=sys.argv[3]
 upload_rates(unix_time, base, quote)
